Point型をdoubleに変更して、座標系の表現をPointに一括化

+++ Position/Point.h
@@ -6,9 +6,9 @@ class Point {
 private:
 	// なし
 public:
-	int X;  // X座標
-	int Y;  // Y座標
-    Point(int X, int Y) : X(X),Y(Y) {}
+	double X;  // X座標
+	double Y;  // Y座標
+    Point(double X, double Y) : X(X),Y(Y) {}
     Point() : Point(0,0) {}
 };
